Refactor : FeignClient package
