+++ src/main/java/com/spacestar/back/feignClient/controller/AuthClient.java
@@ -1,4 +1,4 @@
-package com.spacestar.back.feignClient;
+package com.spacestar.back.feignClient.controller;
 
 import com.spacestar.back.feignClient.dto.res.AuthResDto;
 import org.springframework.cloud.openfeign.FeignClient;
